Added preprocessing, adjusted thresholds
Now the periphery of the image doesn't goof with things nearly as much!
We still have problems with like, not picking out white blobs as
vessels, but things are gradually moving in the right direction.

There's also a bit of code in here that prevents using vesselness in
pixels above a certain intensity.

big todos:
- Thing that automatically records scores for all the images after
trying a set of thresholds
- Saving the useful images generated as a set
- Try a few different thresholds to get this working nice
- Try adding something to remove white splotches