Added code for saving images and outputting stats
All the images are spat out in that /saved folder, and at the end of the
run, stats are printed for each of the images.
Also slightly adjusted the constants B, C, and a threshold for how dark
a pixel has to be for it to be allowed to be a vessel.
Current stats:
 Average accuracy: 0.9351355371900827
 Median accuracy:  0.9377709563164109
 Std accuracy:     0.01905126531412419
 Min accuracy:     0.8934545454545455 (im0003)
 Max accuracy:     0.9699315230224321 (im0291)

 Average sensitivity: 0.5901420421080462
 Median sensitivity:  0.6290558818543878
 Std sensitivity:     0.13037300571507948
 Min sensitivity:     0.325783600653889 (im0324)
 Max sensitivity:     0.7813332240750605 (im0003)

 Average specificity: 0.9770121519467322
 Median specificity:  0.984334781185668
 Std specificity:     0.022961065894476974
 Min specificity:     0.9003114562269947 (im0003)
 Max specificity:     0.9978100725517949 (im0240)